--- run.sh
@@ -1,3 +1,0 @@
-rm a.out
-gcc generate.c ../libft/ft_split.c ../gnl/get_next_line.c ../gnl/get_next_line_utils.c
-./a.out
